Prompt user for deleting invalid save file
Rather than doing it unconditionally. This is a long followup to
834e2b4, which tried to do this in another way and caused big problems.
Hopefully this is the proper way to do it, and it won't cause problems on
each game save since it only happens when validation fails.

+++ src/files.c
@@ -1181,7 +1181,8 @@ restore_saved_game()
     if (validate(fd, fq_save) != 0) {
         (void) nhclose(fd), fd = -1;
 
-        (void) delete_savefile();
+        if (yn("Delete the save file?") == 'y')
+            (void) delete_savefile();
     }
     return fd;
 }
